+++ kernel/data-pipeline/scenario/migration/distsql/statement/src/main/java/org/apache/shardingsphere/data/pipeline/scenario/migration/distsql/statement/updatable/UnregisterMigrationSourceStorageUnitStatement.java
@@ -18,6 +18,7 @@
 package org.apache.shardingsphere.data.pipeline.scenario.migration.distsql.statement.updatable;
 
 import lombok.Getter;
+import lombok.RequiredArgsConstructor;
 import org.apache.shardingsphere.data.pipeline.distsql.statement.updatable.UpdatablePipelineRALStatement;
 
 import java.util.Collection;
@@ -25,12 +26,9 @@ import java.util.Collection;
 /**
  * Unregister migration source storage unit statement.
  */
+@RequiredArgsConstructor
 @Getter
 public final class UnregisterMigrationSourceStorageUnitStatement extends UpdatablePipelineRALStatement {
     
     private final Collection<String> names;
-    
-    public UnregisterMigrationSourceStorageUnitStatement(final Collection<String> names) {
-        this.names = names;
-    }
 }