+++ proxy/frontend/dialect/firebird/src/main/java/org/apache/shardingsphere/proxy/frontend/firebird/command/query/transaction/FirebirdStartTransactionCommandExecutor.java
@@ -39,8 +39,8 @@ public final class FirebirdStartTransactionCommandExecutor implements CommandExe
     
     @Override
     public Collection<DatabasePacket> execute() {
-        connectionSession.setAutoCommit(packet.getAutocommit());
-        connectionSession.setReadOnly(packet.getReadOnly());
+        connectionSession.setAutoCommit(packet.isAutoCommit());
+        connectionSession.setReadOnly(packet.isReadOnly());
         connectionSession.setIsolationLevel(packet.getIsolationLevel());
         int transactionId = FirebirdTransactionIdGenerator.getInstance().nextTransactionId(connectionSession.getConnectionId());
         return Collections.singleton(new FirebirdGenericResponsePacket().setHandle(transactionId));